+++ code/Modules/Gfx/d3d12/d3d12Renderer.cc
@@ -706,7 +706,7 @@ d3d12Renderer::applyDrawState(pipeline* pip, mesh** meshes, int numMeshes) {
 
 //------------------------------------------------------------------------------
 void
-d3d12Renderer::applyUniformBlock(ShaderStage::Code bindStage, int bindSlot, int64_t layoutHash, const uint8_t* ptr, int byteSize) {
+d3d12Renderer::applyUniformBlock(ShaderStage::Code bindStage, int bindSlot, uint32_t layoutHash, const uint8_t* ptr, int byteSize) {
     if (nullptr == this->curPipeline) {
         return;
     }
