+++ code/Modules/Gfx/gl/glRenderer.h
@@ -63,7 +63,7 @@ public:
     /// apply draw state
     void applyDrawState(pipeline* pip, mesh** meshes, int numMeshes, bool record=true);
     /// apply a shader uniform block (called after applyDrawState)
-    void applyUniformBlock(ShaderStage::Code bindStage, int bindSlot, int64_t layoutHash, const uint8_t* ptr, int byteSize, bool record=true);
+    void applyUniformBlock(ShaderStage::Code bindStage, int bindSlot, uint32_t layoutHash, const uint8_t* ptr, int byteSize, bool record=true);
     /// apply a group of textures
     void applyTextures(ShaderStage::Code bindStage, texture** textures, int numTextures, bool record=true);
